Add leaderboard component and page
This commit adds a new leaderboard component and page to the application. The leaderboard component is responsible for displaying a table of users and their scores, while the leaderboard page fetches the leaderboard data from the server and renders the leaderboard component. This feature will allow users to see the rankings and scores of other users in the application.